style: organize and comment the code
