+++ src/storage/storage.cpp
@@ -1,9 +1,323 @@
 #include "storage.hpp"
 
-#include <atomic>
-#include <mutex>
-#include <string>
+#include <dirent.h>
+
+#include <cerrno>
+#include <cstdio>
+#include <cstdlib>
+#include <cstring>
+
+#include <unistd.h>
+
+#include "data_types.hpp"
+#include "esp_log.h"
+#include "esp_system.h"
+#include "esp_vfs.h"
 
 std::atomic<Storage *> Storage::instance;
 std::mutex             Storage::instance_mutex;
 std::string            Storage::logger_tag;
+
+Storage *Storage::getInstance() {
+  ESP_LOGD("Storage", "Acquiring instance...");
+
+  Storage *storage_instance = instance.load(std::memory_order_acquire);
+  if(!storage_instance) {
+    ESP_LOGD("Storage", "Instance does not exist, creating...");
+
+    std::lock_guard<std::mutex> instance_lock(instance_mutex);
+    storage_instance = instance.load(std::memory_order_relaxed);
+    if(!storage_instance) {
+      storage_instance = new Storage();
+      instance.store(storage_instance, std::memory_order_release);
+    }
+  }
+
+  return storage_instance;
+}
+
+bool Storage::is_mounted() {
+  std::lock_guard<std::mutex> lock(fs_mutex);
+  return mount_status;
+}
+
+esp_err_t Storage::mount_storage(std::string mount_path) {
+  std::lock_guard<std::mutex> lock(fs_mutex);
+  const std::string normalized =
+      mount_path.at(0) == '/' ? mount_path : "/" + mount_path;
+  if(mount_status) {
+    if(mount_point == normalized) {
+      return ESP_OK;
+    }
+    ESP_LOGE(logger_tag.c_str(),
+             "FATFS already mounted at %s. To mount at %s, first unmount the "
+             "previous path.",
+             mount_point.c_str(), mount_path.c_str());
+    return ESP_FAIL;
+  }
+
+  mount_point = normalized;
+
+  esp_err_t mount_result = esp_vfs_fat_spiflash_mount_rw_wl(
+      mount_point.c_str(), "storage", &fat_mount_config, &wear_leveling_handle);
+  if(mount_result != ESP_OK) {
+    ESP_LOGE(logger_tag.c_str(), "Failed to mount FATFS (%s)",
+             esp_err_to_name(mount_result));
+    mount_status = false;
+    return ESP_FAIL;
+  }
+  ESP_LOGI(logger_tag.c_str(), "FATFS mounted successfully at %s",
+           mount_point.c_str());
+  mount_status = true;
+  return ESP_OK;
+}
+
+esp_err_t Storage::list_files() {
+  std::lock_guard<std::mutex> lock(fs_mutex);
+  if(!mount_status) {
+    ESP_LOGD(logger_tag.c_str(), "FATFS not mounted.");
+    return ESP_FAIL;
+  }
+
+  DIR           *directory_handle;
+  struct dirent *directory_entry;
+
+  directory_handle = opendir(mount_point.c_str());
+  if(directory_handle) {
+    ESP_LOGD(logger_tag.c_str(), "Listing directory %s:", mount_point.c_str());
+    while((directory_entry = readdir(directory_handle)) != NULL) {
+      ESP_LOGI(logger_tag.c_str(), "File: %s", directory_entry->d_name);
+    }
+    closedir(directory_handle);
+  }
+
+  return ESP_OK;
+}
+
+esp_err_t Storage::save_data(std::string file_path, char *data_buffer,
+                             size_t data_size, const char *mode) {
+  std::lock_guard<std::mutex> lock(fs_mutex);
+  if(!mount_status) {
+    ESP_LOGD(logger_tag.c_str(), "FATFS not mounted.");
+    return ESP_FAIL;
+  }
+
+  std::string full_path   = mount_point + "/" + file_path;
+  FILE       *file_handle = fopen(full_path.c_str(), mode);
+
+  if(file_handle == NULL) {
+    ESP_LOGE(logger_tag.c_str(),
+             "Failed to open file %s for writing: %s (%s)", file_path.c_str(),
+             full_path.c_str(), strerror(errno));
+    return ESP_FAIL;
+  }
+
+  fwrite(data_buffer, data_size, 1, file_handle);
+  fflush(file_handle);
+  int fd = fileno(file_handle);
+  if(fd >= 0) {
+    fsync(fd);
+  }
+  ESP_LOGD(logger_tag.c_str(), "Written %s, %zu bytes", file_path.c_str(),
+           data_size);
+
+  fclose(file_handle);
+
+  return ESP_OK;
+}
+
+esp_err_t Storage::load_data(std::string file_path, char *data_buffer,
+                            size_t data_size) {
+  std::lock_guard<std::mutex> lock(fs_mutex);
+  if(!mount_status) {
+    ESP_LOGD(logger_tag.c_str(), "FATFS not mounted.");
+    return ESP_FAIL;
+  }
+
+  std::string full_path   = mount_point + "/" + file_path;
+  FILE       *file_handle = fopen(full_path.c_str(), "r");
+
+  if(file_handle == NULL) {
+    ESP_LOGE(logger_tag.c_str(),
+             "Failed to open file %s for reading: %s (%s)", file_path.c_str(),
+             full_path.c_str(), strerror(errno));
+    return ESP_FAIL;
+  }
+
+  fread(data_buffer, data_size, 1, file_handle);
+
+  ESP_LOGD(logger_tag.c_str(), "Read %s, %zu bytes", file_path.c_str(),
+           data_size);
+
+  fclose(file_handle);
+
+  return ESP_OK;
+}
+
+esp_err_t Storage::load_data(std::string file_path, char **data_buffer,
+                            size_t *file_size) {
+  std::lock_guard<std::mutex> lock(fs_mutex);
+  if(!mount_status) {
+    ESP_LOGD(logger_tag.c_str(), "FATFS not mounted.");
+    return ESP_FAIL;
+  }
+
+  std::string full_path   = mount_point + "/" + file_path;
+  FILE       *file_handle = fopen(full_path.c_str(), "rb");
+
+  if(file_handle == NULL) {
+    ESP_LOGE(logger_tag.c_str(),
+             "Failed to open file %s for reading: %s (%s)", file_path.c_str(),
+             full_path.c_str(), strerror(errno));
+    return ESP_FAIL;
+  }
+
+  fseek(file_handle, 0, SEEK_END);
+  (*file_size) = ftell(file_handle);
+
+  free(*data_buffer);
+  *data_buffer = (char *)malloc((*file_size) * sizeof(char));
+
+  fseek(file_handle, 0, SEEK_SET);
+
+  fread(*data_buffer, *file_size, 1, file_handle);
+
+  ESP_LOGD(logger_tag.c_str(), "Read %s, %zu bytes", file_path.c_str(),
+           *file_size);
+
+  fclose(file_handle);
+
+  return ESP_OK;
+}
+
+esp_err_t Storage::delete_data(std::string file_path) {
+  std::lock_guard<std::mutex> lock(fs_mutex);
+  if(!mount_status) {
+    ESP_LOGD(logger_tag.c_str(), "FATFS not mounted.");
+    return ESP_FAIL;
+  }
+
+  std::string full_path = mount_point + "/" + file_path;
+  if(remove(full_path.c_str()) != 0) {
+    ESP_LOGE(logger_tag.c_str(), "Failed to remove file %s",
+             file_path.c_str());
+    return ESP_FAIL;
+  }
+
+  ESP_LOGI(logger_tag.c_str(), "Removed %s", file_path.c_str());
+  return ESP_OK;
+}
+
+bool Storage::file_exists(std::string file_path) {
+  std::lock_guard<std::mutex> lock(fs_mutex);
+  if(!mount_status) {
+    return false;
+  }
+
+  std::string full_path   = mount_point + "/" + file_path;
+  FILE       *file_handle = fopen(full_path.c_str(), "r");
+  if(file_handle != NULL) {
+    fclose(file_handle);
+    return true;
+  }
+  return false;
+}
+
+Storage::Storage() {
+  wear_leveling_handle = WL_INVALID_HANDLE;
+  mount_point          = "/data";
+  logger_tag           = "Storage";
+
+  fat_mount_config.format_if_mount_failed = true;
+  fat_mount_config.max_files              = 16;
+  fat_mount_config.allocation_unit_size   = CONFIG_WL_SECTOR_SIZE;
+  mount_status                            = false;
+}
+
+template <typename DataType>
+esp_err_t Storage::write(DataType data_value, std::string file_path) {
+  return save_data(file_path, reinterpret_cast<char *>(&data_value),
+                   sizeof(DataType));
+}
+
+template <typename DataType>
+esp_err_t Storage::read(DataType &data_value, std::string file_path) {
+  return load_data(file_path, reinterpret_cast<char *>(&data_value),
+                   sizeof(DataType));
+}
+
+template <typename ElementType>
+esp_err_t Storage::write_vector(const std::vector<ElementType> &vector_data,
+                               std::string file_path) {
+  std::lock_guard<std::mutex> lock(fs_mutex);
+  if(!mount_status) {
+    ESP_LOGD(logger_tag.c_str(), "FATFS not mounted.");
+    return ESP_FAIL;
+  }
+
+  std::string full_path   = mount_point + "/" + file_path;
+  FILE       *file_handle = fopen(full_path.c_str(), "wb");
+
+  if(file_handle == NULL) {
+    ESP_LOGE(logger_tag.c_str(),
+             "Failed to open file %s for writing: %s (%s)", file_path.c_str(),
+             full_path.c_str(), strerror(errno));
+    return ESP_FAIL;
+  }
+
+  size_t vector_size = vector_data.size();
+  fwrite(&vector_size, sizeof(size_t), 1, file_handle);
+  if(vector_size > 0) {
+    fwrite(vector_data.data(), sizeof(ElementType), vector_size, file_handle);
+  }
+
+  ESP_LOGD(logger_tag.c_str(), "Written vector %s, %zu elements",
+           file_path.c_str(), vector_size);
+
+  fclose(file_handle);
+  return ESP_OK;
+}
+
+template <typename ElementType>
+esp_err_t Storage::read_vector(std::vector<ElementType> &vector_data,
+                               std::string file_path) {
+  std::lock_guard<std::mutex> lock(fs_mutex);
+  if(!mount_status) {
+    ESP_LOGD(logger_tag.c_str(), "FATFS not mounted.");
+    return ESP_FAIL;
+  }
+
+  std::string full_path   = mount_point + "/" + file_path;
+  FILE       *file_handle = fopen(full_path.c_str(), "rb");
+
+  if(file_handle == NULL) {
+    ESP_LOGD(logger_tag.c_str(), "File %s does not exist, using defaults",
+             file_path.c_str());
+    return ESP_FAIL;
+  }
+
+  size_t vector_size = 0;
+  fread(&vector_size, sizeof(size_t), 1, file_handle);
+
+  if(vector_size > 0) {
+    vector_data.resize(vector_size);
+    fread(vector_data.data(), sizeof(ElementType), vector_size, file_handle);
+  } else {
+    vector_data.clear();
+  }
+
+  ESP_LOGD(logger_tag.c_str(), "Read vector %s, %zu elements",
+           file_path.c_str(), vector_size);
+
+  fclose(file_handle);
+  return ESP_OK;
+}
+
+template esp_err_t Storage::write<ParametersConfig>(ParametersConfig,
+                                                    std::string);
+template esp_err_t Storage::read<ParametersConfig>(ParametersConfig &,
+                                                   std::string);
+template esp_err_t Storage::write_vector<MapPoint>(
+    const std::vector<MapPoint> &, std::string);
+template esp_err_t Storage::read_vector<MapPoint>(std::vector<MapPoint> &,
+                                                  std::string);
