--- src/tasks/cli/cli_storage.hpp
@@ -1,5 +1,0 @@
-#pragma once
-
-/// FAT 8.3: basename must be <= 8 chars when LFN is disabled.
-#define MAP_STORAGE_FILE       "map_data.dat"
-#define PARAMETERS_CONFIG_FILE "params.dat"
